LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/reporting/access-reporting/access-from-coop-page-to-openee_coop-ro.https-expected.txt


FAIL access-from-coop-page-to-openee, same-origin assert_equals: expected (string) "coop" but got (undefined) undefined
FAIL access-from-coop-page-to-openee, same-origin + redirect assert_equals: expected (string) "coop" but got (undefined) undefined
